+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<a name="v0.68.2"></a>
+## [v0.68.2] - 2021-12-03
+### Bug Fixes
+- **entities:** handle deprecated field errors in tests
+- **http:** check if the NerdGraph error is a deprecation warning, and still pass on the response (with error)
+
 <a name="v0.68.1"></a>
 ## [v0.68.1] - 2021-11-29
 ### Bug Fixes
@@ -1210,7 +1216,8 @@
 - extract paging implementation
 - rename packages for clarity, promote Config to the public package
 
-[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.68.1...HEAD
+[Unreleased]: https://github.com/newrelic/newrelic-client-go/compare/v0.68.2...HEAD
+[v0.68.2]: https://github.com/newrelic/newrelic-client-go/compare/v0.68.1...v0.68.2
 [v0.68.1]: https://github.com/newrelic/newrelic-client-go/compare/v0.68.0...v0.68.1
 [v0.68.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.67.0...v0.68.0
 [v0.67.0]: https://github.com/newrelic/newrelic-client-go/compare/v0.66.2...v0.67.0